--- gnn/configs/T5.json
@@ -1,57 +1,0 @@
-{
-    "basedir": "/blue/p.chang/jguiang/data/lst/GATOR/CMSSW_12_2_0_pre2",
-    "name": "QuintupletClassifier",
-    "model": {
-        "name": "ChangNet",
-        "n_hidden_layers": 4,
-        "hidden_size": 200,
-        "n_message_features": 10,
-        "n_message_passing_rounds": 3,
-        "message_aggregator": "sum"
-    },
-    "ingress": {
-        "input_file": "/blue/p.chang/jguiang/data/lst/GATOR/CMSSW_12_2_0_pre2/LSTGnnNtuple_noT5Chi2_noPixelsInTCs.root",
-        "ttree_name": "tree", 
-        "undirected": false,
-        "edge_indices": [
-            "t5_t3_idx0",
-            "t5_t3_idx1"
-        ],
-        "edge_features": [
-            "t5_pt",
-            "t5_eta",
-            "t5_phi",
-            "t5_chiSquared",
-            "t5_innerRadius",
-            "t5_bridgeRadius",
-            "t5_outerRadius"
-        ],
-        "node_features": [
-            "t3_pt",
-            "t3_eta",
-            "t3_phi",
-            "t3_0_z",
-            "t3_0_r",
-            "t3_0_dr",
-            "t3_0_layer",
-            "t3_2_z",
-            "t3_2_r",
-            "t3_2_dr",
-            "t3_2_layer",
-            "t3_4_z",
-            "t3_4_r",
-            "t3_4_dr",
-            "t3_4_layer"
-        ],
-        "truth_label": "t5_isFake",
-        "branch_filter": "/(t3|t5)_*/"
-    },
-    "train": {
-        "train_batch_size": 64,
-        "test_batch_size": 1000,
-        "learning_rate": 0.002,
-        "learning_rate_step_gamma": 0.8,
-        "learning_rate_step_size": 25,
-        "seed": 1234
-    }
-}
